Refactored core code
